Enhanced storybook redux injection
