+++ tests/storybook/utils/connectForm.js
@@ -1,0 +1,12 @@
+import { reduxForm } from "redux-form";
+import { connect } from "react-redux";
+
+const connectForm = FormComponent => {
+  const connected = connect()(FormComponent);
+  const Form = reduxForm({
+    form: "storybook",
+  })(connected);
+  return Form;
+};
+
+export default connectForm;
